requirements.txt
aiohttp
requests
python-dotenv
